error found working on fix
