+++ GLAPS/FrontEnd/FrontEnd/FrontEnd/__init__.py
@@ -8,7 +8,4 @@ app = Flask(__name__)
 from FrontEnd import db #seems to be working
 db.init_app(app)
 
-from . import auth
-app.register_blueprint(auth.bp)
-
 import FrontEnd.views
