+++ GLAPS/FrontEnd/FrontEnd/FrontEnd/db.py
@@ -20,7 +20,7 @@ def init_app(app):
 def get_db():
     if 'db' not in g:
         g.db = sqlite3.connect(
-            current_app.config['DATABASE'],
+            current_app.config['glaps.sqlite'],
             detect_types=sqlite3.PARSE_DECLTYPES
         )
         g.db.row_factory = sqlite3.Row
